refactor(datenbank): migrate WorkDetail to use child routing
- use RouterView for expression details
- Add child route DatenbankWorkExpressionDetail

RISDEV-8517
